chore(interfaces): move deployment event types to generic file
